fix: refactor service to retire futs
