+++ examples/hello-world-native/hello-index-native/src/main.rs
@@ -27,11 +27,13 @@ mod hello_world_native {
     async fn index_logged_greeting(event: Greeting, block: BlockData) {
         // Since all events require a u64 ID field, let's derive an ID using the
         // name of the person in the Greeting
-        let greeter_id = first8_bytes_to_u64(&event.person.name.to_string());
+        let greeter_name = trim_sized_ascii_string(&event.person.name);
+        let greeting = trim_sized_ascii_string(&event.greeting);
+        let greeter_id = first8_bytes_to_u64(&greeter_name);
 
         // Here we 'get or create' a Salutation based on the ID of the event
         // emitted in the LogData receipt of our smart contract
-        let greeting = match Salutation::load(event.id).await {
+        let salutation = match Salutation::load(event.id).await {
             Some(mut g) => {
                 // If we found an event, let's use block height as a proxy for time
                 g.last_seen = block.height;
@@ -40,8 +42,7 @@ mod hello_world_native {
             None => {
                 // If we did not already have this Saluation stored in the database. Here we
                 // show how you can use the Charfield type to store strings with length <= 255
-                let message =
-                    format!("{} 👋, my name is {}", &event.greeting, &event.person.name);
+                let message = format!("{} 👋, my name is {}", &greeting, &greeter_name);
 
                 Salutation {
                     id: event.id,
@@ -65,13 +66,17 @@ mod hello_world_native {
             None => Greeter {
                 id: greeter_id,
                 first_seen: block.height,
-                name: event.person.name.to_string(),
+                name: greeter_name,
                 last_seen: block.height,
+
+                // Here we show an example of an arbtrarily sized Blob type. These Blob types
+                // support data up to 10485760 bytes in length
+                visits: vec![1u8, 2, 3, 4, 5, 6, 7, 8],
             },
         };
 
         // Both entity saves will occur in the same transaction
-        greeting.save().await;
+        salutation.save().await;
         greeter.save().await;
     }
 }
