+++ examples/hello-world-native/schema/hello_index_native.schema.graphql
@@ -13,6 +13,7 @@ type Greeter {
     name: Charfield!
     first_seen: UInt8!
     last_seen: UInt8!
+    visits: Blob!
 }
 
 # Calling this `Salutation` so as to not clash with `Greeting` in the contract
