+++ packages/fuel-indexer-macros/src/indexer.rs
@@ -777,11 +777,8 @@ pub fn process_indexer_module(attrs: TokenStream, item: TokenStream) -> TokenStr
                 service.register_native_index(manifest, handle_events).await?;
                 let service_handle = tokio::spawn(service.run());
 
-                if cfg!(feature = "api-server") {
-                    let _ = tokio::join!(service_handle, GraphQlApi::build_and_run(config, pool, tx));
-                } else {
-                    service_handle.await?;
-                };
+                // FIXME: should still respect feature flags
+                let _ = tokio::join!(service_handle, GraphQlApi::build_and_run(config, pool, tx));
 
                 Ok(())
             }
