+++ packages/fuel-indexer/Cargo.toml
@@ -31,6 +31,7 @@ futures = "0.3"
 http = "0.2"
 hyper = { version = "0.14", features = ["client", "http2", "http1", "runtime" ]}
 hyper-rustls = { version = "0.23", features = ["http2"] }
+lazy_static = "1.4"
 prometheus = { version = "0.13", optional = true }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = { version = "1.0", features = ["raw_value"] }
